data

+++ src/TaskSeven.java
@@ -3,18 +3,18 @@ public class TaskSeven {
     public static void main(String[] args) {
         Scanner enter = new Scanner(System.in);
         System.out.print("Введите число: ");
-        double a = enter.nextDouble();
-        double b = (int) (a / 100);
-        double c = a - b * 100 ;
-        double f = (int) (c / 10) ;
-        double e = c - 10 * f ;
+        int a = enter.nextInt();
+        int b =  a / 100;
+        int c = a - b * 100 ;
+        int f =  (c / 10) ;
+        int e = c - 10 * f ;
         if ( e == 1 && c != 11) {
-            System.out.println((int )a + " Програмист");
+            System.out.println(a + " Програмист");
         } else {
             if (e > 1 && e < 5 && c != 12 && c != 13 && c != 14 ){
-                System.out.println((int) a + " Програмиста");
+                System.out.println( a + " Програмиста");
             }else{
-                System.out.println((int) a + " Програмистов");
+                System.out.println( a + " Програмистов");
             }
         }
             }
